backend/test_ingest.py
import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Now import and test
import json

def test_ingestion():
    """Test the complete ingestion pipeline with Elasticsearch."""
    print("Testing ingestion pipeline...")

    # Sample repo - using your actual repository for demonstration
    repo_url = "https://github.com/elipaulman/GOVS"  # Your actual repo

    try:
        print(f"Starting ingestion of {repo_url}...")

        # Test Elasticsearch connection first
        try:
            es = get_elasticsearch_client()
            es.ping()
            print("✅ Elasticsearch connection successful")
        except Exception as e:
            print(f"⚠️  Elasticsearch not available ({e})")
            print("   This is expected if Elasticsearch is not running locally")
            print("   The ingestion will still work but data won't be indexed")
            return False

        # Check if index exists and create if needed
        index_name = "repo_chunks"
        if es.indices.exists(index=index_name):
            print(f"✅ Index '{index_name}' already exists")
        else:
            print(f"📦 Creating index '{index_name}'...")
            es.indices.create(index=index_name, body={
                "mappings": {
                    "properties": {
                        "repo_owner": {"type": "keyword"},
                        "repo_name": {"type": "keyword"},
                        "file_path": {"type": "keyword"},
                        "content": {"type": "text"},
                        "embedding": {"type": "dense_vector", "dims": 768},
                        "chunk_id": {"type": "keyword"},
                        "timestamp": {"type": "date"}
                    }
                }
            })
            print(f"✅ Created index '{index_name}'")

        # Ingest the repository
        ingest_github_repo(repo_url)
        print("✅ Ingestion completed successfully.")

        # Verify data was indexed
        count_result = es.count(index=index_name)
        doc_count = count_result["count"]
        print(f"✅ Indexed {doc_count} chunks in Elasticsearch")

        return True
    except Exception as e:
        print(f"❌ Error during ingestion: {e}")
        import traceback
        traceback.print_exc()
        return False

def test_search():
    """Test the search functionality."""
    print("\nTesting search functionality...")

    try:
        # Test Elasticsearch connection first
        try:
            es = get_elasticsearch_client()
            es.ping()
            print("✅ Elasticsearch connection successful")
        except Exception as e:
            print(f"⚠️  Elasticsearch not available ({e}), creating mock data for testing")
            return test_search_with_mock_data()

        # Search for some common programming terms
        test_queries = [
            "function",
            "class",
            "import",
            "README"
        ]

        for query in test_queries:
            print(f"\nSearching for: '{query}'")
            results = search_similar_chunks(query, top_k=3)

            if results:
                print(f"✅ Found {len(results)} results:")
                for i, result in enumerate(results, 1):
                    print(f"  {i}. {result['file_path']} (score: {result['score']:.3f})")
                    print(f"     Content: {result['content'][:100]}...")
            else:
                print("  No results found")

        return True
    except Exception as e:
        print(f"❌ Error during search: {e}")
        import traceback
        traceback.print_exc()
        return False

def test_search_with_mock_data():
    """Test search functionality with mock data when Elasticsearch is not available."""
    print("\n🔧 Testing with mock data...")

    try:
        # Create some mock search results
        mock_results = [
            {
                "content": "# Hello World\n\nThis is a sample README file for testing the ingestion pipeline.",
                "file_path": "README.md",
                "repo_name": "Hello-World",
                "repo_owner": "octocat",
                "metadata": {"language": "markdown"},
                "score": 0.892,
                "chunk_id": "abc123"
            },
            {
                "content": "def hello_world():\n    print('Hello, World!')\n    return True",
                "file_path": "hello.py",
                "repo_name": "Hello-World",
                "repo_owner": "octocat",
                "metadata": {"language": "python"},
                "score": 0.847,
                "chunk_id": "def456"
            },
            {
                "content": "class Greeter:\n    def greet(self, name):\n        return f'Hello, {name}!'",
                "file_path": "greeter.py",
                "repo_name": "Hello-World",
                "repo_owner": "octocat",
                "metadata": {"language": "python"},
                "score": 0.723,
                "chunk_id": "ghi789"
            }
        ]

        test_queries = ["function", "class", "README"]

        for query in test_queries:
            print(f"\nSearching for: '{query}'")
            # Filter mock results based on query
            filtered_results = [r for r in mock_results if query.lower() in r['content'].lower()]

            if filtered_results:
                print(f"✅ Found {len(filtered_results)} results:")
                for i, result in enumerate(filtered_results, 1):
                    print(f"  {i}. {result['file_path']} (score: {result['score']:.3f})")
                    print(f"     Content: {result['content'][:100]}...")
            else:
                print("  No results found")

        return True
    except Exception as e:
        print(f"❌ Error during mock search test: {e}")
        return False

def test_basic_functionality():
    """Test basic functionality without requiring full pipeline."""
    print("\n🧪 Testing basic functionality...")

    try:
        # Test GitHub API integration
        from github_utils import get_repo_files, get_file_content

        print("Testing GitHub API integration...")
        files = get_repo_files("elipaulman", "GOVS")
        print(f"✅ Found {len(files)} files in GOVS repo")

        # Test file content retrieval
        readme_content = get_file_content("elipaulman", "GOVS", "README.md")
        print(f"✅ Retrieved README.md ({len(readme_content)} characters)")

        # Test text splitting
        from langchain.text_splitter import RecursiveCharacterTextSplitter

        print("Testing text splitting...")
        try:
            # Simple text splitting without TextLoader for testing
            splitter = RecursiveCharacterTextSplitter(chunk_size=1000, chunk_overlap=100)
            chunks = splitter.split_text(readme_content)
            print(f"✅ Split into {len(chunks)} chunks")
        except Exception as e:
            print(f"⚠️  Text splitting test failed: {e}")
            # Fallback: simple character-based splitting
            chunk_size = 1000
            chunks = []
            for i in range(0, len(readme_content), chunk_size):
                chunk = readme_content[i:i + chunk_size]
                chunks.append(chunk)
            print(f"✅ Fallback split into {len(chunks)} chunks")

        return True
    except Exception as e:
        print(f"❌ Error in basic functionality test: {e}")
        return False

if __name__ == "__main__":
    print("🚀 Repo Rover - Complete Ingestion Pipeline Test")
    print("=" * 50)

    # Test basic functionality first
    basic_success = test_basic_functionality()

    if basic_success:
        print("\n✅ Basic functionality test passed!")
        # Test ingestion
        ingestion_success = test_ingestion()

        # Test search (only if ingestion was successful)
        if ingestion_success:
            test_search()
        else:
            print("\n⚠️  Skipping search test due to ingestion failure")
            # Still try search with mock data
            test_search_with_mock_data()
    else:
        print("\n⚠️  Basic functionality test failed, running mock tests only")
        test_search_with_mock_data()

    print("\n" + "=" * 50)
    print("Test completed!")
